Adopt the UIScene life cycle
Apple deprecated the app-delegate-only life cycle with the iOS 26 SDK and
has said it will stop working in a future release. The app still had a
UIWindow on AppDelegate and no scene manifest at all.

Both Info.plists get a UIApplicationSceneManifest pointing at a new
SceneDelegate, and UIMainStoryboardFile is replaced by the manifest's
UISceneStoryboardFile. Window ownership and the document opening path move
from AppDelegate to SceneDelegate, split across willConnectTo for a cold
launch and openURLContexts for an already running app. The Inbox move
keeps working the same way, minus a force unwrap on the documents
directory.

UIApplicationSupportsMultipleScenes stays false. Turning it on is what
would give iPad multi window, but DocumentBrowserViewController keeps
per-instance state that has never been exercised with more than one
scene, so that belongs in its own change.

The six empty applicationWill/Did lifecycle stubs are gone; they only ever
held Xcode's template comments.
