+++ OpenDocumentReader/AppDelegate.swift
@@ -11,10 +11,7 @@ import UIKit
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
-    var window: UIWindow?
-
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        
         // same split OpenDocument.droid uses: reporting only in the ad
         // supported flavor, nothing at all in the paid one
         let reportingEnabled = ConfigurationManager.manager.configuration == .lite
@@ -22,69 +19,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         CrashManager.shared.setEnabled(reportingEnabled)
 
         StoreReviewHelper.incrementAppOpenedCount()
-        
-        return true
-    }
-
-    func applicationWillResignActive(_ application: UIApplication) {
-        // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
-        // Use this method to pause ongoing tasks, disable timers, and invalidate graphics rendering callbacks. Games should use this method to pause the game.
-    }
-
-    func applicationDidEnterBackground(_ application: UIApplication) {
-        // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later.
-        // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
-    }
-
-    func applicationWillEnterForeground(_ application: UIApplication) {
-        // Called as part of the transition from the background to the active state; here you can undo many of the changes made on entering the background.
-    }
-
-    func applicationDidBecomeActive(_ application: UIApplication) {
-        // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
-    }
-
-    func applicationWillTerminate(_ application: UIApplication) {
-        // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
-    }
-
-    func application(_ app: UIApplication, open inputURL: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
-        guard let documentBrowserViewController = window?.rootViewController as? DocumentBrowserViewController else {
-            CrashManager.shared.log("root view controller is not a DocumentBrowserViewController")
-
-            return false
-        }
-        
-        let documentsUrl = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first!
-        let inboxUrl = documentsUrl.appendingPathComponent("Inbox")
-
-        let destinationUrl: URL
-        if (inputURL.absoluteString.contains(inboxUrl.absoluteString)) {
-            do {
-                destinationUrl = documentsUrl.appendingPathComponent(inputURL.lastPathComponent)
-
-                try FileManager.default.moveItem(at: inputURL, to: destinationUrl)
-            } catch {
-                CrashManager.shared.log(error)
-
-                return false
-            }
-        } else {
-            destinationUrl = inputURL
-        }
-        
-        documentBrowserViewController.revealDocument(at: destinationUrl, importIfNeeded: true) { (revealedDocumentURL, error) in
-            guard let documentUrl = revealedDocumentURL else {
-                // ignoring errors because they should pop up in failedToImportDocumentAt too
-                
-                return;
-            }
-            
-            documentBrowserViewController.presentDocument(at: documentUrl)
-        }
 
         return true
     }
-    
+
+    func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
+        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
+    }
 }
-
